requirements.txt
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
scikit-image==0.18.3
fastapi==0.68.0
Pillow==8.3.1
joblib==1.1.0
uvicorn==0.15.0
